FlaskRestAPIWithCloudTask/requirements.txt
Flask
Flask-RESTful
gunicorn==20.1.0
google-cloud-tasks==2.10.4